+++ craco.config.js
@@ -52,6 +52,9 @@ module.exports = {
         filename: 'bundle.[name].js',
         path: path.resolve(__dirname, 'build'),
       };
+      webpackConfig.externals = {
+        'fsevents': "require('fsevents')",
+      };
       // Temporary solution until react-scripts 5.0.1 is released
       webpackConfig.ignoreWarnings = [/Failed to parse source map/];
       return webpackConfig;