+++ src/components/organisms/ExplorerPane/ExplorerPane.tsx
@@ -2,6 +2,10 @@ import {Collapse as RawCollapse} from 'antd';
 
 import styled from 'styled-components';
 
+import {isInClusterModeSelector} from '@redux/appConfig';
+import {useAppDispatch, useAppSelector} from '@redux/hooks';
+import {setExplorerSelectedSection} from '@redux/reducers/ui';
+
 import {trackEvent} from '@shared/utils/telemetry';
 
 import FilePane from './FilePane';
@@ -10,13 +14,21 @@ import ImagesPane from './ImagesPane';
 import KustomizePane from './KustomizePane';
 
 const ExplorerPane: React.FC = () => {
+  const dispatch = useAppDispatch();
+  const explorerSelectedSection = useAppSelector(state => state.ui.explorerSelectedSection);
+  const isInClusterMode = useAppSelector(isInClusterModeSelector);
+
   return (
     <Collapse
       accordion
       ghost
-      defaultActiveKey="files"
-      onChange={key => {
-        trackEvent('left-menu/activity-changed', {activity: 'explorer', section: key as string});
+      activeKey={isInClusterMode ? 'images' : explorerSelectedSection}
+      onChange={(key: any) => {
+        if (isInClusterMode) {
+          return;
+        }
+        dispatch(setExplorerSelectedSection(key));
+        trackEvent('left-menu/activity-changed', {activity: 'explorer', section: key});
       }}
     >
       <FilePane key="files" />