+++ src/components/organisms/ExplorerPane/FilePane/FileSystemTree/TreeNode/hooks.tsx
@@ -10,7 +10,12 @@ import {basename, dirname, join, sep} from 'path';
 import {scanExcludesSelector, updateProjectConfig} from '@redux/appConfig';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setAlert} from '@redux/reducers/alert';
-import {openCreateFileFolderModal, openNewResourceWizard, openRenameEntityModal} from '@redux/reducers/ui';
+import {
+  openCreateFileFolderModal,
+  openFileCompareModal,
+  openNewResourceWizard,
+  openRenameEntityModal,
+} from '@redux/reducers/ui';
 import {useResourceMetaMapRef} from '@redux/selectors/resourceMapSelectors';
 import {getLocalResourceMetasForPath} from '@redux/services/fileEntry';
 import {getHelmValuesFile, isHelmChartFile, isHelmTemplateFile, isHelmValuesFile} from '@redux/services/helm';
@@ -22,6 +27,7 @@ import {deleteFileEntry, dispatchDeleteAlert, duplicateEntity, isFileEntryDisabl
 import {useOpenOnGithub} from '@utils/git';
 import {useRefSelector} from '@utils/hooks';
 
+import {isYamlFile} from '@monokle/validation';
 import {AlertEnum} from '@shared/models/alert';
 import {FileEntry} from '@shared/models/fileEntry';
 import {isDefined} from '@shared/utils/filter';
@@ -418,6 +424,16 @@ export const useFileMenuItems = (
         duplicate(fileEntry);
       },
     });
+
+    if (isYamlFile(fileEntry.filePath)) {
+      newMenuItems.push({
+        key: 'compare',
+        label: 'Compare with another file',
+        onClick: () => {
+          dispatch(openFileCompareModal(fileEntry.filePath));
+        },
+      });
+    }
 
     newMenuItems.push(...commonMenuItems);