+++ src/components/organisms/ValidationPane/ValidationPane.tsx
@@ -3,7 +3,8 @@ import {useMeasure} from 'react-use';
 import {Image} from 'antd';
 import Link from 'antd/lib/typography/Link';
 
-import {useAppDispatch} from '@redux/hooks';
+import {isInClusterModeSelector} from '@redux/appConfig';
+import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setLeftMenuSelection} from '@redux/reducers/ui';
 import {useValidationSelector} from '@redux/validation/validation.selectors';
 import {setSelectedProblem} from '@redux/validation/validation.slice';
@@ -12,13 +13,14 @@ import {usePaneHeight} from '@hooks/usePaneHeight';
 
 import ValidationFigure from '@assets/NewValidationFigure.svg';
 
-import {TitleBar, ValidationOverview} from '@monokle/components';
+import {Icon, TitleBar, ValidationOverview} from '@monokle/components';
 
 import * as S from './ValidationPane.styled';
 
 const ValidationPane: React.FC = () => {
   const dispatch = useAppDispatch();
   const lastResponse = useValidationSelector(state => state.lastResponse);
+  const isInClusterMode = useAppSelector(isInClusterModeSelector);
   const newProblemsIntroducedType = useValidationSelector(state => state.validationOverview.newProblemsIntroducedType);
   const selectedProblem = useValidationSelector(state => state.validationOverview.selectedProblem);
   const status = useValidationSelector(state => state.status);
@@ -35,7 +37,7 @@ const ValidationPane: React.FC = () => {
     <S.ValidationPaneContainer>
       <div ref={titleBarRef}>
         <TitleBar
-          title="Validation Overview"
+          title={`Validation Overview ${isInClusterMode ? '( Cluster mode )' : ''}`}
           description={
             <S.DescriptionContainer>
               <Image src={ValidationFigure} width={95} />
@@ -43,6 +45,15 @@ const ValidationPane: React.FC = () => {
                 Fix your resources according to your validation setup. Manage your validation policy, turn rules on or
                 off, and more in the <Link onClick={() => dispatch(setLeftMenuSelection('settings'))}>settings</Link>{' '}
                 section, located in the left menu.
+                {isInClusterMode && (
+                  <S.BackToDashboardButton
+                    icon={<Icon name="cluster-dashboard" />}
+                    type="primary"
+                    onClick={() => dispatch(setLeftMenuSelection('dashboard'))}
+                  >
+                    Back to cluster dashboard
+                  </S.BackToDashboardButton>
+                )}
               </div>
             </S.DescriptionContainer>
           }
@@ -51,6 +62,7 @@ const ValidationPane: React.FC = () => {
 
       <ValidationOverview
         containerStyle={{marginTop: '20px'}}
+        showOnlyByResource={isInClusterMode}
         height={height - titleBarHeight - 40}
         newProblemsIntroducedType={newProblemsIntroducedType}
         selectedProblem={selectedProblem?.problem}