+++ src/redux/reducers/ui.ts
@@ -1,20 +1,23 @@
 import {webFrame} from 'electron';
 
-import {Draft, PayloadAction, createSlice} from '@reduxjs/toolkit';
+import {Draft, PayloadAction, createSlice, isAnyOf} from '@reduxjs/toolkit';
 
 import path from 'path';
 import {Entries} from 'type-fest';
 
 import {DEFAULT_PANE_CONFIGURATION} from '@constants/constants';
 
+import {activeProjectSelector} from '@redux/appConfig';
 import initialState from '@redux/initialState';
-import {loadClusterResources} from '@redux/thunks/cluster';
+import {AppListenerFn} from '@redux/listeners/base';
+import {loadClusterResources, stopClusterConnection} from '@redux/thunks/cluster';
 import {setRootFolder} from '@redux/thunks/setRootFolder';
 
 import {ROOT_FILE_ENTRY} from '@shared/constants/fileEntry';
 import {Project, SavedCommand, SettingsPanel} from '@shared/models/config';
 import {ResourceIdentifier} from '@shared/models/k8sResource';
 import {
+  ExplorerCollapsibleSectionsType,
   HighlightItems,
   LayoutSizeType,
   LearnTopicType,
@@ -315,6 +318,18 @@ export const uiSlice = createSlice({
     closeKeyboardShortcutsModal: (state: Draft<UiState>) => {
       state.isKeyboardShortcutsModalOpen = false;
     },
+    openFileCompareModal: (state: Draft<UiState>, action: PayloadAction<string>) => {
+      state.fileCompareModal = {
+        isVisible: true,
+        filePath: action.payload,
+      };
+    },
+    closeFileCompareModal: (state: Draft<UiState>) => {
+      state.fileCompareModal = {
+        isVisible: false,
+        filePath: '',
+      };
+    },
     openScaleModal: (state: Draft<UiState>) => {
       state.isScaleModalOpen = true;
     },
@@ -351,6 +366,12 @@ export const uiSlice = createSlice({
     searchHelmRepo: (state: Draft<UiState>, action: PayloadAction<string>) => {
       state.helmRepo.search = action.payload;
     },
+    setExplorerSelectedSection: (state: Draft<UiState>, action: PayloadAction<ExplorerCollapsibleSectionsType>) => {
+      state.explorerSelectedSection = action.payload;
+    },
+    setFileExplorerExpandedFolders: (state: Draft<UiState>, action: PayloadAction<string[]>) => {
+      state.fileExplorerExpandedFolders = action.payload;
+    },
   },
   extraReducers: builder => {
     builder
@@ -382,6 +403,7 @@ export const {
   closeAboutModal,
   closeCreateFileFolderModal,
   closeCreateProjectModal,
+  closeFileCompareModal,
   closeFiltersPresetModal,
   closeFolderExplorer,
   closeKeyboardShortcutsModal,
@@ -401,6 +423,7 @@ export const {
   openAboutModal,
   openCreateFileFolderModal,
   openCreateProjectModal,
+  openFileCompareModal,
   openFiltersPresetModal,
   openFolderExplorer,
   openKeyboardShortcutsModal,
@@ -418,6 +441,8 @@ export const {
   setActiveSettingsPanel,
   setExpandedFolders,
   setExpandedSearchedFiles,
+  setExplorerSelectedSection,
+  setFileExplorerExpandedFolders,
   setLayoutSize,
   setLeftBottomMenuSelection,
   setLeftMenuIsActive,
@@ -450,3 +475,24 @@ export const {
   setHelmRepoPane,
 } = uiSlice.actions;
 export default uiSlice.reducer;
+
+// Listeners
+
+export const stopClusterConnectionListener: AppListenerFn = listen => {
+  listen({
+    matcher: isAnyOf(stopClusterConnection.fulfilled),
+    effect: async (action, {getState, dispatch}) => {
+      const activeProject = activeProjectSelector(getState());
+
+      if (activeProject) {
+        return;
+      }
+
+      const leftMenuSelection = getState().ui.leftMenu.selection;
+
+      if (leftMenuSelection === 'validation') {
+        dispatch(setLeftMenuSelection('dashboard'));
+      }
+    },
+  });
+};